Switch to non-sparse multiplication for AVX2. This now turns out to be faster on my Intel laptop. To compile with sparse multiplication, add sparse=yes to the make command. To compile without sparse multiplication, add sparse=no. The default is sparse=yes, except for AVX2, BMI2 and AVX512.
I suspect that sparse=yes is still faster on Zen 1 and perhaps also on some other CPUs with AVX2. This needs to be tested.
